+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Renamed `--no-builtin-keywords` to `--disable-builtin-keywords`
 - Added a --pager flag to override the default pager in addition to the env variable
 - Fixed a bug where --listen-command would never run the pager
+- Renamed `--listen-command` to `--exec`
 
 ## 5.3.0